LayoutTests/accessibility/mac/scrollbars.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head>
<body id="body" role="group">

<script>
    description("This tests that a scroll area will return scroll bars when asked.");

    // Cause the web page to gain scrollbars.
    for (var k = 0; k < 100; k++) {
       document.getElementById("body").innerHTML += "<br><br><br>";
    }

    var scrollArea;
    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        setTimeout(async function() {
            // Iterate up from the body AX element to the scrollarea AX element.
            let maybeScrollArea = (await waitForElementById("body")).parentElement();
            await waitFor(() => {
                if (maybeScrollArea && maybeScrollArea.role === "AXRole: AXScrollArea") {
                    scrollArea = maybeScrollArea;
                    return true;
                }

                maybeScrollArea = maybeScrollArea.parentElement();
                return false;
            });

            shouldBe("scrollArea.role", "'AXRole: AXScrollArea'");
            shouldBe("scrollArea.verticalScrollbar.role", "'AXRole: AXScrollBar'");

            finishJSTest();
        }, 0);
    }
</script>
</body>
</html>
